use double instead of SMB_BIG_UINT for dir_total and ttarf (by tridge, merged from branch) (This used to be commit b482ac8fc4948ead79be8dd08c8386449f12e5f6)

+++ source3/client/clitar.c
@@ -1487,7 +1487,7 @@ int process_tar(void)
     free(tarbuf);
     
     DEBUG(0, ("tar: dumped %d tar files\n", ntarf));
-    DEBUG(0, ("Total bytes written: %llu\n", ttarf));
+    DEBUG(0, ("Total bytes written: %.0f\n", (double)ttarf));
     break;
   }
